Change from user to web and added showtoken, passwd
